Refactor aspen.resources
This collapses the resources/ directory down to a single resources.py
module, which exposes Dynamic and Static classes (formerly
SimplateWrapper and StaticResource) and, more importantly, `get` and
`load` functions to return instances of Dynamic and Static (get caches,
load doesn't). The test_unbound_simplate.py test script is updated as
well.
