+++ tests/test_dynamic_resource.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 from pytest import raises, yield_fixture
 
 from aspen import resources, Response
-from aspen.resources import SimplateWrapper
+from aspen.resources import Dynamic
 from aspen.simplates.pagination import Page
 from aspen.simplates.renderers.stdlib_template import Factory as TemplateFactory
 from aspen.simplates.renderers.stdlib_percent import Factory as PercentFactory
@@ -21,17 +21,17 @@ def get(harness):
                  , default_media_type = ''
                   )
         kw.update(_kw)
-        return SimplateWrapper(**kw)
+        return Dynamic(**kw)
     yield get
 
 
-def test_unbound_simplate_is_instantiable(harness):
+def test_dynamic_resource_is_instantiable(harness):
     website = harness.client.website
     fs = ''
     raw = '[---]\n[---] text/plain via stdlib_template\n'
     media_type = ''
-    actual = SimplateWrapper(website, fs, raw, media_type).__class__
-    assert actual is SimplateWrapper
+    actual = Dynamic(website, fs, raw, media_type).__class__
+    assert actual is Dynamic
 
 
 # compile_page
@@ -106,7 +106,7 @@ def test_get_renderer_factory_can_raise_syntax_error(get):
     assert msg.startswith("Malformed renderer oo*gle. It must match")
 
 
-# get_response
+# respond
 
 def _get_state(harness, *a, **kw):
     kw['return_after'] = 'dispatch_request_to_filesystem'
@@ -114,13 +114,13 @@ def _get_state(harness, *a, **kw):
     return harness.simple(*a, **kw)
 
 
-def _get_response(state):
+def _respond(state):
     resource = resources.load(state['website'], state['dispatch_result'].match, 0)
     state['resource'] = resource
     state['response'] = state.get('response', Response())
     return resource.respond(state)
 
-UNBOUND_SIMPLATE = """\
+SIMPLATE = """\
 [---]
 [---] text/plain
 Greetings, program!
@@ -128,83 +128,83 @@ Greetings, program!
 <h1>Greetings, program!</h1>
 """
 
-def test_get_response_gets_response(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
+def test_respond_responds(harness):
+    harness.fs.www.mk(('index.spt', SIMPLATE))
     response = Response()
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     state['response'] = response
-    actual = _get_response(state)
+    actual = _respond(state)
     assert actual is response
 
-def test_get_response_is_happy_not_to_negotiate(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
-    actual = _get_response(state).body
+def test_respond_is_happy_not_to_negotiate(harness):
+    harness.fs.www.mk(('index.spt', SIMPLATE))
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
+    actual = _respond(state).body
     assert actual == "Greetings, program!\n"
 
-def test_get_response_sets_content_type_when_it_doesnt_negotiate(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
-    actual = _get_response(state).headers['Content-Type']
+def test_respond_sets_content_type_when_it_doesnt_negotiate(harness):
+    harness.fs.www.mk(('index.spt', SIMPLATE))
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
+    actual = _respond(state).headers['Content-Type']
     assert actual == "text/plain; charset=UTF-8"
 
-def test_get_response_doesnt_reset_content_type_when_not_negotiating(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+def test_respond_doesnt_reset_content_type_when_not_negotiating(harness):
+    harness.fs.www.mk(('index.spt', SIMPLATE))
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     response = Response()
     response.headers['Content-Type'] = 'never/mind'
     state['response'] = response
-    actual = _get_response(state).headers['Content-Type']
+    actual = _respond(state).headers['Content-Type']
     assert actual == "never/mind"
 
-def test_get_response_negotiates(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+def test_respond_negotiates(harness):
+    harness.fs.www.mk(('index.spt', SIMPLATE))
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     state['accept_header'] = 'text/html'
-    actual = _get_response(state).body
+    actual = _respond(state).body
     assert actual == "<h1>Greetings, program!</h1>\n"
 
 def test_handles_busted_accept(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+    harness.fs.www.mk(('index.spt', SIMPLATE))
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     # Set an invalid Accept header so it will return default (text/plain)
     state['request'].headers['Accept'] = 'text/html;'
-    actual = _get_response(state).body
+    actual = _respond(state).body
     assert actual == "Greetings, program!\n"
 
-def test_get_response_sets_content_type_when_it_negotiates(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+def test_respond_sets_content_type_when_it_negotiates(harness):
+    harness.fs.www.mk(('index.spt', SIMPLATE))
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     state['accept_header'] = 'text/html'
-    actual = _get_response(state).headers['Content-Type']
+    actual = _respond(state).headers['Content-Type']
     assert actual == "text/html; charset=UTF-8"
 
-def test_get_response_doesnt_reset_content_type_when_negotiating(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+def test_respond_doesnt_reset_content_type_when_negotiating(harness):
+    harness.fs.www.mk(('index.spt', SIMPLATE))
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     state['request'].headers['Accept'] = 'text/html'
     response = Response()
     response.headers['Content-Type'] = 'never/mind'
     state['response'] = response
-    actual = _get_response(state).headers['Content-Type']
+    actual = _respond(state).headers['Content-Type']
     response = Response()
     response.headers['Content-Type'] = 'never/mind'
     state['response'] = response
-    actual = _get_response(state).headers['Content-Type']
+    actual = _respond(state).headers['Content-Type']
     assert actual == "never/mind"
 
-def test_get_response_raises_406_if_need_be(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+def test_respond_raises_406_if_need_be(harness):
+    harness.fs.www.mk(('index.spt', SIMPLATE))
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     state['accept_header'] = 'cheese/head'
-    actual = raises(Response, _get_response, state).value.code
+    actual = raises(Response, _respond, state).value.code
     assert actual == 406
 
-def test_get_response_406_gives_list_of_acceptable_types(harness):
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE))
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+def test_respond_406_gives_list_of_acceptable_types(harness):
+    harness.fs.www.mk(('index.spt', SIMPLATE))
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     state['accept_header'] = 'cheese/head'
-    actual = raises(Response, _get_response, state).value.body
+    actual = raises(Response, _respond, state).value.body
     expected = "The following media types are available: text/plain, text/html."
     assert actual == expected
 
@@ -224,23 +224,23 @@ def install_glubber(harness):
 
 def test_can_override_default_renderers_by_mimetype(harness):
     install_glubber(harness)
-    harness.fs.www.mk(('index.spt', UNBOUND_SIMPLATE),)
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+    harness.fs.www.mk(('index.spt', SIMPLATE),)
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     state['request'].headers['Accept'] = 'text/plain'
-    actual = _get_response(state).body
+    actual = _respond(state).body
     assert actual == "glubber"
 
 def test_can_override_default_renderer_entirely(harness):
     install_glubber(harness)
-    state = _get_state(harness, filepath='index.spt', contents=UNBOUND_SIMPLATE)
+    state = _get_state(harness, filepath='index.spt', contents=SIMPLATE)
     state['request'].headers['Accept'] = 'text/plain'
-    actual = _get_response(state).body
+    actual = _respond(state).body
     assert actual == "glubber"
 
 
 # indirect
 
-INDIRECTLY_NEGOTIATED_UNBOUND_SIMPLATE = """\
+INDIRECTLY_NEGOTIATED_SIMPLATE = """\
 [-------]
 foo = "program"
 [-------] text/html
@@ -249,26 +249,26 @@ foo = "program"
 Greetings, %(foo)s!"""
 
 def test_indirect_negotiation_sets_media_type(harness):
-    harness.fs.www.mk(('/foo.spt', INDIRECTLY_NEGOTIATED_UNBOUND_SIMPLATE))
+    harness.fs.www.mk(('/foo.spt', INDIRECTLY_NEGOTIATED_SIMPLATE))
     response = harness.client.GET('/foo.html')
     expected = "<h1>Greetings, program!</h1>\n"
     actual = response.body
     assert actual == expected
 
 def test_indirect_negotiation_sets_media_type_to_secondary(harness):
-    harness.fs.www.mk(('/foo.spt', INDIRECTLY_NEGOTIATED_UNBOUND_SIMPLATE))
+    harness.fs.www.mk(('/foo.spt', INDIRECTLY_NEGOTIATED_SIMPLATE))
     response = harness.client.GET('/foo.txt')
     expected = "Greetings, program!"
     actual = response.body
     assert actual == expected
 
 def test_indirect_negotiation_with_unsupported_media_type_is_404(harness):
-    harness.fs.www.mk(('/foo.spt', INDIRECTLY_NEGOTIATED_UNBOUND_SIMPLATE))
+    harness.fs.www.mk(('/foo.spt', INDIRECTLY_NEGOTIATED_SIMPLATE))
     response = harness.client.GxT('/foo.jpg')
     assert response.code == 404
 
 
-UNBOUND_SIMPLATE_VIRTUAL_PATH = """\
+SIMPLATE_VIRTUAL_PATH = """\
 [-------]
 foo = request.path['foo']
 [-------] text/html
@@ -277,14 +277,14 @@ foo = request.path['foo']
 Greetings, %(foo)s!"""
 
 
-def test_unbound_inside_virtual_path(harness):
-    harness.fs.www.mk(('/%foo/bar.spt', UNBOUND_SIMPLATE_VIRTUAL_PATH ))
+def test_dynamic_resource_inside_virtual_path(harness):
+    harness.fs.www.mk(('/%foo/bar.spt', SIMPLATE_VIRTUAL_PATH ))
     response = harness.client.GET('/program/bar.txt')
     expected = "Greetings, program!"
     actual = response.body
     assert actual == expected
 
-UNBOUND_SIMPLATE_STARTYPE = """\
+SIMPLATE_STARTYPE = """\
 [-------]
 foo = request.path['foo']
 [-------] */*
@@ -294,21 +294,21 @@ Unknown request type, %(foo)s!
 [-------] text/*
 Greetings, %(foo)s!"""
 
-def test_unbound_inside_virtual_path_with_startypes_present(harness):
-    harness.fs.www.mk(('/%foo/bar.spt', UNBOUND_SIMPLATE_STARTYPE ))
+def test_dynamic_resource_inside_virtual_path_with_startypes_present(harness):
+    harness.fs.www.mk(('/%foo/bar.spt', SIMPLATE_STARTYPE ))
     response = harness.client.GET('/program/bar.html')
     actual = response.body
     assert '<h1>' in actual
 
-def test_unbound_inside_virtual_path_with_startype_partial_match(harness):
-    harness.fs.www.mk(('/%foo/bar.spt', UNBOUND_SIMPLATE_STARTYPE ))
+def test_dynamic_resource_inside_virtual_path_with_startype_partial_match(harness):
+    harness.fs.www.mk(('/%foo/bar.spt', SIMPLATE_STARTYPE ))
     response = harness.client.GET('/program/bar.txt')
     expected = "Greetings, program!"
     actual = response.body
     assert actual == expected
 
-def test_unbound_inside_virtual_path_with_startype_fallback(harness):
-    harness.fs.www.mk(('/%foo/bar.spt', UNBOUND_SIMPLATE_STARTYPE ))
+def test_dynamic_resource_inside_virtual_path_with_startype_fallback(harness):
+    harness.fs.www.mk(('/%foo/bar.spt', SIMPLATE_STARTYPE ))
     response = harness.client.GET('/program/bar.jpg')
     expected = "Unknown request type, program!"
     actual = response.body.strip()
